# Source/server/swagger.yaml
swagger: "2.0"
info:
  description: "API của Website Hỗ trợ đào tạo SGU"
  version: "1.0.0"
  title: "Website Hỗ trợ đào tạo SGU"
  contact:
    email: ""
host: ""
...
basePath: "/api"
schemes:
- "http"
- "https"
...
paths:
...
  /admin/khoa-luan-tot-nghiep/ThemDeTaiKhoaLuan/{MaKLTN}:
    post:
      tags:
        - "KhoaLuanTotNghiep"
      summary: "Thêm đề tài khóa luận tốt nghiệp"
      produces:
        - application/json
      parameters:
        - in: path
          name: MaKLTN
          type: string
          required: true
          description: Mã khóa luận tốt nghiệp
        - in: body
          name: TenDeTai
          type: string
          required: true
          description: Tên đề tài
        - in: body
          name: MaGV
          type: string
          required: true
          description: Mã giảng viên
      responses:
        200:
          description: Thêm đề tài khóa luận tốt nghiệp thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Thêm đề tài khóa luận tốt nghiệp thành công.",
            }
        400:
          description: Thêm đề tài khóa luận tốt nghiệp thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message" : "Thêm đề tài khóa luận tốt nghiệp thất bại.",
            }
      security:
        - ApiKeyAuth: []
  /admin/khoa-luan-tot-nghiep/ChinhSuaTenDeTaiKhoaLuan/{MaKLTN}:
    post:
      tags:
        - "KhoaLuanTotNghiep"
      summary: "Chỉnh sửa tên đề tài khóa luận tốt nghiệp"
      produces:
        - application/json
      parameters:
        - in: path
          name: MaKLTN
          type: string
          required: true
          description: Mã khóa luận tốt nghiệp
        - in: body
          name: TenDeTaiCu
          type: string
          required: true
          description: Tên đề tài cũ
        - in: body
          name: TenDeTaiMoi
          type: string
          required: true
          description: Tên đề tài mới
        - in: body
          name: MaGV
          type: string
          required: true
          description: Mã giảng viên
      responses:
        200:
          description: Chỉnh sửa tên đề tài khóa luận tốt nghiệp thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Chỉnh sửa tên đề tài khóa luận tốt nghiệp thành công.",
            }
        400:
          description: Chỉnh sửa tên đề tài khóa luận tốt nghiệp thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message" : "Chỉnh sửa tên đề tài khóa luận tốt nghiệp thất bại.",
            }
      security:
        - ApiKeyAuth: []
  /admin/khoa-luan-tot-nghiep/XoaDeTaiKhoaLuan/{MaKLTN}:
    post:
      tags:
        - "KhoaLuanTotNghiep"
      summary: "Xóa đề tài khóa luận tốt nghiệp"
      produces:
        - application/json
      parameters:
        - in: path
          name: MaKLTN
          type: string
          required: true
          description: Mã khóa luận tốt nghiệp
        - in: body
          name: TenDeTai
          type: string
          required: true
          description: Tên đề tài
        - in: body
          name: MaGV
          type: string
          required: true
          description: Mã giảng viên
      responses:
        200:
          description: Xóa đề tài khóa luận tốt nghiệp thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Xóa đề tài khóa luận tốt nghiệp thành công.",
            }
        400:
          description: Xóa đề tài khóa luận tốt nghiệp thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message" : "Xóa đề tài khóa luận tốt nghiệp thất bại.",
            }
      security:
        - ApiKeyAuth: []
...
  /khoa-luan-tot-nghiep/ThongTinKLTN:
    get:
      tags:
        - "KhoaLuanTotNghiepClient"
      summary: "Xem chi tiết thông tin khóa luận tốt nghiệp"
      produces:
        - application/json
      responses:
        200:
          description: Xem chi tiết thông tin khóa luận tốt nghiệp thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Xem chi tiết thông tin khóa luận tốt nghiệp thành công.",
            }
        400:
          description: Xem chi tiết thông tin khóa luận tốt nghiệp thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message" : "Xem chi tiết thông tin khóa luận tốt nghiệp thất bại.",
            }
      security:
        - ApiKeyAuth: []
  /khoa-luan-tot-nghiep/DSDeTaiTheoGiangVien/{MaKLTN}:
    post:
      tags:
        - "KhoaLuanTotNghiepClient"
      summary: "Lấy danh sách đề tài của giảng viên"
      produces:
        - application/json
      parameters:
        - in: path
          name: MaKLTN
          type: string
          required: true
          description: Mã khóa luận tốt nghiệp
        - in: body
          name: MaGV
          type: string
          required: true
          description: Mã giảng viên
      responses:
        200:
          description: Lấy danh sách đề tài của giảng viên thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Lấy danh sách đề tài của giảng viên thành công.",
            }
        400:
          description: Lấy danh sách đề tài của giảng viên thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message" : "Lấy danh sách đề tài của giảng viên thất bại.",
            }
      security:
        - ApiKeyAuth: []
  /khoa-luan-tot-nghiep/SVDangKyDeTai/{MaKLTN}:
    post:
      tags:
        - "KhoaLuanTotNghiepClient"
      summary: "Sinh viên đăng ký đề tài khóa luận"
      produces:
        - application/json
      parameters:
        - in: path
          name: MaKLTN
          type: string
          required: true
          description: Mã khóa luận tốt nghiệp
        - in: body
          name: TenDeTai
          type: string
          required: true
          description: Tên đề tài
        - in: body
          name: MaGV
          type: string
          required: true
          description: Mã giảng viên
        - in: body
          name: MaSV
          type: string
          required: true
          description: Mã sinh viên
        - in: body
          name: HoSV
          type: string
          required: true
          description: Họ sinh viên
        - in: body
          name: TenSV
          type: string
          required: true
          description: Tên sinh viên
        - in: body
          name: Email
          type: string
          required: true
          description: Email sinh viên
        - in: body
          name: SoDienThoai
          type: string
          required: true
          description: Số điện thoại sinh viên
        - in: body
          name: DTBTL
          type: string
          required: true
          description: Điểm trung bình tích lũy học kỳ (hệ 4)
        - in: body
          name: TinChiTL
          type: string
          required: true
          description: Số tín chỉ tích lũy
      responses:
        200:
          description: Sinh viên đăng ký đề tài khóa luận thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Sinh viên đăng ký đề tài khóa luận thành công.",
            }
        400:
          description: Sinh viên đăng ký đề tài khóa luận thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message" : "Sinh viên đăng ký đề tài khóa luận thất bại.",
            }
      security:
        - ApiKeyAuth: []
...
  /khoa-luan-tot-nghiep/GVXoaSVDangKyKLDuKien/{MaKLTN}:
    post:
      tags:
        - "KhoaLuanTotNghiepClient"
      summary: "Xóa thông tin sinh viên dự kiến đăng ký khóa luận tốt nghiệp"
      produces:
        - application/json
      parameters:
        - in: path
          name: MaKLTN
          type: string
          required: true
          description: Mã khóa luận tốt nghiệp
        - in: body
          name: TenDeTai
          type: string
          required: true
          description: Tên đề tài
        - in: body
          name: MaGV
          type: string
          required: true
          description: Mã giảng viên
        - in: body
          name: MaSV
          type: string
          required: true
          description: Mã sinh viên
      responses:
        200:
          description: Xóa thông tin sinh viên dự kiến đăng ký khóa luận tốt nghiệp thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Xóa thông tin sinh viên dự kiến đăng ký khóa luận tốt nghiệp thành công.",
            }
        400:
          description: Xóa thông tin sinh viên dự kiến đăng ký khóa luận tốt nghiệp thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message" : "Xóa thông tin sinh viên dự kiến đăng ký khóa luận tốt nghiệp thất bại.",
            }
      security:
        - ApiKeyAuth: []
...
  /khoa-luan-tot-nghiep/DSDeTaiChuaDangKy/{MaKLTN}:
    post:
      tags:
        - "KhoaLuanTotNghiepClient"
      summary: "Lấy danh sách đề tài chưa đăng ký"
      produces:
        - application/json
      parameters:
        - in: path
          name: MaKLTN
          type: string
          required: true
          description: Mã khóa luận tốt nghiệp
      responses:
        200:
          description: Lấy danh sách đề tài chưa đăng ký thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Lấy danh sách đề tài chưa đăng ký thành công.",
            }
        400:
          description: Lấy danh sách đề tài chưa đăng ký thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message" : "Lấy danh sách đề tài chưa đăng ký thất bại.",
            }
      security:
        - ApiKeyAuth: []
...
securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: headers
    name: x-access-token